requirements.txt
watchdog
pdfplumber
python-dotenv